l10n_fr,l10n_jp: Fix translation of "Open"
In English, "Open" means both the action of openning and the state when something
is open. Several workflows have the problem that the "Open" transition is
translated to something that does not describe the action, but the state.

We have a mechanism to override the translation for a state, by translating
`Word [{state_title} state in {workflow_id}]`, but we don't have such mechanism
for transitions, so when "Word" exist both as a transition and a state, "Word"
should be the translation of the transition and we should have translations in
context for all the workflow states.