ERP5Type/tests/Sequence: be quiet by default
Sequence produce lots of useless output, which makes it slow to display test
result pages for tests using sequences, like testBusinesTemplate.
It must also be taking a lot of disk space on the server and is generally
not useful, when a test fail we have the log of the failing test.