payment_mean(_sepa): generate reference when grouping
For SEPA files "End to End ID" we need a reference on payment transactions.
In the first prototype, the reference was set at the time of generating the
file, but it's a non common behaviour to generate reference when producing
reports, so we changed to ensuring payment has reference when adding to the
group.