mimetypes_registry: merge image/x-ms-bmp in Windows BMP image
We had two mimetypes entries, which caused inconsistencies depending on
wether the lookup was done by mimetype, by glob or by extension.

We had:

- name:
   "Windows BMP image"
- mimetypes:
    image/bmp
    image/x-bmp
    image/x-MS-bmp
- extensions:
- globs:
    *.bmp

and

- name:
   "image/x-ms-bmp"
- mimetypes:
    image/x-ms-bmp
- extensions:
    bmp
- globs:

With this commit they are merged into one:

- name:
   "Windows BMP image"
- mimetypes:
    image/x-ms-bmp
    image/bmp
    image/x-bmp
    image/x-MS-bmp
- extensions:
    bmp
- globs:
    *.bmp

This way we only have one consistent mimetype.

For compatibility with extension lookups (that are done in
Document_guessMimeType interaction workflow from DMS), image/x-ms-bmp is
kept as default. This might not be the best choice, according to
https://www.iana.org/assignments/media-types/media-types.xhtml