core_test: Add test to make sure that wendelin.core basically works
Wendelin.core is now integral part of ERP5 (see [1,2]), but nothing
inside ERP5 currently uses it. And even though wendelin.core has its own
testsuite, integration problems are always possible.

-> Add test to erp5_core_test that minimally makes sure that basic
wendelin.core operations work.

This test currently passes with wendelin.core 1, which is the default.
It also passes as live test with wendelin.core 2.
However with wendelin.core 2 it currently fails on testnodes like e.g.

    ValueError: ZODB.MappingStorage.MappingStorage is in-RAM storage
	in-RAM storages are not supported:
	a zurl pointing to in-RAM storage in one process would lead to
	another in-RAM storage in WCFS process.

and

    RuntimeError: wcfs: join file:///srv/slapgrid/slappart8/srv/testnode/djk/test_suite/unit_test.2/var/Data.fs: server not started
    (https://nexedijs.erp5.net/#/test_result_module/20210530-92EF3124/102)

because we need to amend ERP5 test driver

1) to run tests on a real storage instead of in-RAM Mapping Storage(*), and
2) to spawn WCFS server for each such storage.

I will try to address those points in a later patch.

In the meantime there should be no reason not to merge this, because we
do not use wendelin.core 2 yet, and solving "1" and "2" first are
preconditions to begin such a usage.

/cc @rafael, @tomo, @seb, @jerome, @romain, @vpelletier, @Tyagov, @klaus, @jp

(*) Combining Zope and WCFS working together requires data to be on a real
    storage, not on in-RAM MappingStorage inside Zope's Python process.

[1] https://lab.nexedi.com/nexedi/slapos/commit/7f877621
[2] https://lab.nexedi.com/nexedi/slapos/merge_requests/874#note_122339