pdm: fix Organisation_jumpToRelatedObjectList
getRelatedValueList seems to have changed in behavior at some point, when
passed multiple base_category_list, it only returns document that are member
of all the categories. Since this API is inneficient here (it searchs catalog
unrestrictedly and then apply check one by one the documents in python), we
change to use portal_catalog which is much better suited for this kind of
queries.