ZODB Components: add getContentType method
This is mostly for editor fields, to fix the problem that components were
edited by the text editor and not the code editor.

Editor is selected by Base_getEditorFieldPreferredTextEditor, which will
use a text editor by default, unless the document has a content type, in
which case it will return the most appropriate editor, which is by default
the source code editor for documents whose text content is not text/html.

It's anyway not wrong to have a content type defined here, Components
are TextDocuments and TextDocuments usually have getContentType method.