ERP5Form: fix `cell` in Matrixbox TALES context
TALESValue always had a strange behavior of setting the request as `cell` when
rendering a matrixbox in the case where there was no cell at the given
coordinates, but this code was not used, becase the request was not passed in
kw. When in 777c5e6cbb ([ERP5Form] Pass request to fields, 2017-12-29) we
changed to always pass the request, this changed the behavior. After this
change when there was no matrix cell, in the TALES context the variable `cell`
became request.

This restore the original behavior (cell being the cell or None), by changing
the questionable code that was not used before.