CaptchaField: enable haproxy sticky cookie when viewing a captcha
Currently portal_sessions is only in RAM, so captcha only work when the client
is lucky enough to hit the same zope when submitting the captcha for valiation.

This workaround improves things a little bit because it forces the balancer to
set cookie. Unfortunately, it does not really work because it's too late for
the current request, so basically it will work on the next request.

Unless we can fix portal_sessions, a better workaround is to use an intermediate
script before the page containing the captcha, to set the cookie and redirect
to the page with captcha.