SessionTool: support storing temp objects
This was only supported for the RAMCache version of session tool, but is also
good to support them in DistributedRamCache, so that we can use DistributedRamCache
by default.

One breaking change is that when using RAMCache, changes made to temp objects
in session were persisting for the session lifetime, after this change, each
time we get the object from the session we get a new fresh object, so if we
make change we have to explicitly put it back to the session, which is surprising
for developers used to the conventience of ZODB transparent persistence.