base: fix Person_getAvailableAssignmentValueList for assignments with only start date
Because stop date gets the start date by acquisition when no stop date is set,
scripts need to take special care that the date is actually set. This was not
the case in this script, so assignments with only a start date but no stop date
where not considered.

This was especially visible in Base_getDialogSectionCategoryItemList which is
used to select the group in user preferences and in reports dialogs