officejs_support_request_ui: more flexible way of making preview link for attachments
Instead of expecting exactly actions with "preview" or "view" name, take any
action which name contains "preview" (or title is "Preview") or name is "view".

This makes it easier to include support request applications in different
applications, where the default actions might have a different name. This also
fixes the preview link for File portal type, because there was no "preview" or
"view" actions for File.