graph_editor: don't include user action transitions in workflow graph
These are rendered with a loop from the state to the state, it's ugly and not
much informative, remove them for now. We could use different edge styles for
"worklow method only" and "workflow method and user action" transitions and
render this as only one transition. We could also change the data model of
workflow itself, now that this is no longer using DCWorkflow.