knowledge_pad: use a better key for messages
To "mark as read" and open preview, the front page gadget uses the "md5" key,
which is calculated from "link", but we can have multiple entries for the same
links, like it's the case for officejs_support_request app.

Instead, use guid or id if available, this is what RSS/atom use for this purpose.