core/mrp: make sure getAggregatedAmountList return amounts in order
For some reports displaying transformations result to end users, it's much
better to have amount ordered in the same order as the model lines.

AmountGeneratorMixin was already sorting the lines, but the order was lost
during aggregation, because we used a dict, which on python2 does not
retain order. Switch to using an OrderedDict to keep the order.