ERP5Type/patches/WSGIPublisher: Consider the error handled if err_hook reraised.
err_hook may re-raise when it is specifically told not to render an error
page. Such exception does not mean err_hook failed, so it should not be
re-raised by WSGIPublisher: re-raising in WSGIPublisher means handing the
exception over to waitress, which will have no other choice than producing a
generic error page that we have no control over.