ERP5Type/patches/WSGIPublisher: Consider the error handled if err_hook reraised.
err_hook may re-raise when it is specifically told not to render an error
page. But such exception does not mean the error hasn't been handled
according to the transaction's desires: such transaction may be an
interface implementation which already set the response up and only need
the transaction to be rolled back.
Re-raising here means handing the exception off to waitress, which will
cook an error page we have no control over. This is not good.
If the error raised by err_hook is not a reraise, pass it on to the
response so it can make an effort at producing a response where err_hook
failed. If this raises, then to waitress we go.