TimerServer: use an int as SERVER_PORT in environment
Since https://github.com/zopefoundation/Zope/commit/2da2c2c28a8b21e78e8ef93aa17da926e3102c26
SERVER_PORT is assumed to be convertible to int.

The value of this port is not really supposed to be used because
TimerServer requests does not have a server (for example in CMFActivity
we recreate the initial request), so put a value of -1 to prevent
accidental use.