zopewsgi: start TimerServer from startup script
When running with ZServer, TimerServer will be started by ZConfig
machinery, because ZServer comes with some ZConfig data types for server
factory which start servers.

One Zope 2, ZServer was imported from zopeschema:

https://github.com/zopefoundation/Zope/blob/2.13.30/src/Zope2/Startup/zopeschema.xml#L10

and ZConfig used the TimerServerFactory to create a TimerService.

On Zope 4, ZServer is not imported anymore in the new wsgischema:

https://github.com/zopefoundation/Zope/blob/4.6.3/src/Zope2/Startup/wsgischema.xml

so the ZConfig data type can not be used to create the server.

Instead of relying on ZServer's ZConfig, we create the server directly
from the startup script, the logic to attach to waitress is already here
and startup and factory just called create, so the only difference is
that we no longer use zope.conf to configure TimerService, but configure
from the script.

SlapOS profile also needs to be modified in order to not include ZConfig
bits for timerserver and to pass the timerserver interval in the wrapper
to this script.