erp5_core: Make Preference_disableOtherPreferences consistent.
Ownership and Owner role are distinct: catalog's "viewable_owner" is
role-based, but "owner" is ownership based. As permissions are applied by
catalog in addition to any given parameter, this should not change the
result of this query. And as any owner which does not have the Owner role
would cause this script to fail on the assertion, this should not change
the result of this script in any way besides avoiding AssertionErrors.