dms: rename print actions not to conflict with erp5_odt_style
odt_style comes with "Print" global actions which allows getting the
default form as ODT and convert it to selected format and DMS comes with
"Print" actions which return a PDF of the document content.

As a result, when user try to use Print, they have the same action twice
with no way to know which one is which.

This is detected by CodingStyleTestCase.test_DuplicateActions, but only
for configurations where both erp5_odt_style and erp5_dms are installed,
which is not the case in coding style test suite which only install the
miminmal dependent business templates.