pdm: support defining function on supply lines
We make it possible to define function the same way it's currently
possible to define accounts. The idea is similar: to be able to use this
information in accounting generation rules.

The typical accounting generation configuration using this will use the
function defined on the supply line if any is defined and with a fallback
to the function defined on business process' trade model path.

For now we don't introduce such configuration by default, but this may
become part of  the configuration generated by configurator some day.