base: better handling of passwords set to None
If by code we create a login with:

    user.newContent(portal_type='ERP5 Login', password=None)

this created a login with password 'None'.

Note that happens for newContent, not for login.edit(password=None),
because newContent calls edit with force_update=1 (see [1] for code
reference), so this probably only happened in this test.

To prevent accidentally setting a password to None when force_update=1
is used, we add the same None check in _setPassword than the one we
already had in setPassword.

[1] https://lab.nexedi.com/nexedi/erp5/blob/ab62760f5321cc015aec613df7c178ec137bc3ce/product/ERP5Type/ERP5Type.py#L448