Products.ERP5Type.Core.Predicate: asQuery should not force an empty result.
When asQuery is used among more conditions (ex: when publishing documents in a
Web Section), the absence of any predicate criterion should not apply any
condition on the caller.
This allows the caller to detect the absence of any condition and to take
the appropriate decision.
Maintain previous behaviour in callers which did not check a None return
value.
Also, extend the effect of isEmptyCriterionValid to the test method, as
it should be consistent with asQuery.