ERP5Site_checkNamingConventions: Accept the use of 'your_' in field libraries.
The only use 'my_' has above 'your_' is to provide a default field value
without needing a TALES expression. This only gets applied based on the
field ID in the form being rendered. Field libraries are never meant to be
rendered, so using 'my_' is always (if harmless) pointless. What really
matters for the field naming convention (which exist to avoid collisions
with form properties) is that *some* prefix is used, be it 'my_' or
'your_'.
So update this check rule to tolerate 'your_' prefixes in addition to
'my_'.
Also, use 'not any([...])' instead of 'not 1 in [...]'.