tests: fix counting selenium failures
1b1dbf601c (tests: also consider python unittest failures in
functional tests, 2021-06-16) was not counting properly the cases
where we have selenium failures. In that case we only want to count
selenium failures, if we add with the python failures from
status_dict, we report one extra failures.

The correct approach is to count selenium failures if any and otherwise
count python failures

Co-authored-by: Vincent Pelletier <vincent@nexedi.com>