corporate_identity_test: relax html snapshot test
Instead of testing that the produced html is byte-to-byte identical
to the expected snapshot, use lxml's doctest support[1] to compare,
so that we can tolerate differences in whitespaces and also order
of attributes.

Also make this test a bit easier to debug by dumping html in the test
log folders.

[1]: https://lxml.de/lxmlhtml.html#running-html-doctests