base: add missing docstrings on classes
Zope only publish objects if they have a docstring. The change in class
hierarchy from b7a604780e (ERP5Security,erp5: simplify
EncryptedPasswordMixin.setPassword, 2022-02-21) caused that Login and
subclasses no longer have a docstring and as a result there were no
longer publishable by Zope, so it was not possible to view them in xhtml
style. ERP5JS was not affected, because it does not directly rely on Zope
publication APIs.